Adapterクラスを別ファイルに切りだし
